Rename `bool` union member
In C23, `bool` has been changed from a macro defined in `<stdbool.h>`, which
expands to `_Bool`, to a keyword.  The `var` union defined in wmgeneral.h
contains a member named `bool`, which is, therefore, not compatible with C23.
The next version of gcc defaults to C23, and so this union member will cause
compilation to fail.  Rename it to `boolean`.

Replace `if (cond) { x = 1; } else { x = 0; }` with `x = cond;`

Signed-off-by: Jeremy Sowden <jeremy@azazel.net>